requirements.txt
fastapi
uvicorn[standard]
sqlalchemy[asyncio]
aiomysql
alembic
pydantic-settings
python-dotenv
langgraph
langchain
langchain-community
langchain-ollama
greenlet
chromadb 
pypdf